# backend/openapi/AsikaGo.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AsikaGo.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/roadmap": {
      "get": {
        "tags": [
          "RoadmapEndpoints"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/roadmap/steps/{number}": {
      "get": {
        "tags": [
          "RoadmapEndpoints"
        ],
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RegistrationStatus": {
        "enum": [
          "Planning",
          "Started"
        ]
      },
      "RequirementResponse": {
        "required": [
          "id",
          "name",
          "description",
          "appliesTo",
          "source"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "appliesTo": {
            "type": [
              "null",
              "string"
            ]
          },
          "source": {
            "$ref": "#/components/schemas/SourceResponse"
          }
        }
      },
      "RoadmapResponse": {
        "required": [
          "roadmapId",
          "createdAt",
          "businessType",
          "categoryName",
          "cityName",
          "registrationStatus",
          "steps"
        ],
        "type": "object",
        "properties": {
          "roadmapId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "businessType": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "cityName": {
            "type": "string"
          },
          "registrationStatus": {
            "$ref": "#/components/schemas/RegistrationStatus"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoadmapStepResponse"
            }
          }
        }
      },
      "RoadmapStepResponse": {
        "required": [
          "number",
          "stepId",
          "name",
          "agency",
          "conditionNote",
          "requirementCount"
        ],
        "type": "object",
        "properties": {
          "number": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "stepId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "agency": {
            "type": "string"
          },
          "conditionNote": {
            "type": [
              "null",
              "string"
            ]
          },
          "requirementCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "SourceResponse": {
        "required": [
          "name",
          "url",
          "dateVerified"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "dateVerified": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "StepDetailResponse": {
        "required": [
          "number",
          "totalSteps",
          "stepId",
          "name",
          "agency",
          "description",
          "actions",
          "conditionNote",
          "requirements",
          "previousNumber",
          "nextNumber"
        ],
        "type": "object",
        "properties": {
          "number": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalSteps": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "stepId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "agency": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "actions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "conditionNote": {
            "type": [
              "null",
              "string"
            ]
          },
          "requirements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequirementResponse"
            }
          },
          "previousNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nextNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      }
    }
  },
...
}